Move resolve_multiple_rewriteable_revsets to cli_utils.rs
